pyproject.toml
[project]
name = "ai-architect-final-srt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "chromadb>=1.5.9",
    "datasets>=4.8.5",
    "httpx>=0.28.1",
    "kagglehub>=1.0.1",
    "langchain>=1.2.18",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.3",
    "langchain-huggingface>=1.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.10",
    "mcp>=1.27.1",
    "openai>=2.36.0",
    "pandas>=3.0.3",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.57.0",
]
